🤖 Merge PR #56506 [react-addons-perf] Update definitions for 15.4 by @eps1lon
